+++ sdk/schemaregistry/schema-registry/recordings/node/schemaregistryclient/recording_sets_endpoint_in_constructor.js
@@ -1,5 +1,0 @@
-let nock = require('nock');
-
-module.exports.hash = "46c11a0b5cbacff73a981772862a4d39";
-
-module.exports.testInfo = {"uniqueName":{},"newDate":{}}