+++ sdk/core/core-client/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added a new function `createCAECallbacks`, that generates callbacks to use on the `bearerTokenAuthenticationPolicy` to support [Continuous Access Evaluation (CAE) challenges](https://docs.microsoft.com/azure/active-directory/conditional-access/concept-continuous-access-evaluation).
+
 ### Breaking Changes
 
 ### Bugs Fixed